# Conflicts:
#	bali-framework-dependencies/pom.xml

+++ bali-auth-server/src/main/java/com/github/bali/auth/service/impl/RoleServiceImpl.java
@@ -63,6 +63,10 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
             log.error("user:{},attempts to remove the super administrator role", SecurityUtil.getUser().getId());
             throw new BaseRuntimeException("警告：你无法删除超级管理员角色");
         }
+        if (SecurityConstant.TENANT_ADMIN_ROLE.equals(role.getRole())) {
+            log.error("user:{},attempts to remove the tenant administrator role", SecurityUtil.getUser().getId());
+            throw new BaseRuntimeException("警告：你无法删除租户管理员角色");
+        }
         userRoleService.remove(Wrappers.<UserRole>lambdaQuery().eq(UserRole::getRoleId, id));
         return this.removeById(id);
     }
@@ -82,8 +86,12 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
     public Boolean update(Role role) {
         Role r = Optional.ofNullable(this.get(role.getId())).orElseGet(Role::new);
         if (SecurityConstant.SUPER_ADMIN_ROLE.equals(r.getRole())) {
-            log.error("user:{},attempts to remove the super administrator role", SecurityUtil.getUser().getId());
+            log.error("user:{},attempts to update the super administrator role", SecurityUtil.getUser().getId());
             throw new BaseRuntimeException("警告：你无法修改超级管理员角色");
+        }
+        if (SecurityConstant.TENANT_ADMIN_ROLE.equals(r.getRole())) {
+            log.error("user:{},attempts to update the tenant administrator role", SecurityUtil.getUser().getId());
+            throw new BaseRuntimeException("警告：你无法修改租户管理员角色");
         }
         Role updateRole = new Role();
         updateRole.setId(role.getId());
